integrations/google_cloud_run_jobs/requirements.txt
google-cloud-run>=0.10.0
dagster>=1.6.0
